made this changes in models.py. set the initialization of MLPs and LSTMs as in the paper. removed sigmoid activation that was applied to the importance score outputted from the LSTM. removed multiplying (the importance score output) by the phi(output from MLP-similarity) and returned the phi directly. in in eval.py, set the dp variable type to float instead of int, which was causing a problem.

+++ evaluation.py
@@ -10,7 +10,7 @@ def knapsack_dp(values, weights, capacity):
     weights = np.array(weights, dtype=np.int32)
     capacity = int(capacity)
     
-    dp = np.zeros((n + 1, capacity + 1), dtype=np.int32)
+    dp = np.zeros((n + 1, capacity + 1), dtype=np.float32)
     
     # Build the DP table
     for i in range(1, n + 1):
